refactor: 템플릿 컴포넌트로 이동
