+++ src/App.ts
@@ -1,55 +1,39 @@
-import addEventBottomSheet from "./components/BottomSheet";
-import BottomSheetButton from "./components/BottomSheetButton";
-import addEventFilterSortButton from "./components/FilterSortButton";
-import { RestaurantList } from "./RestaurantList";
+import BottomSheet from "./components/BottomSheet";
+import BottomSheetOpenButton from "./components/BottomSheetOpenButton";
+import FilterSortButton from "./components/FilterSortButton";
+import { RestaurantService } from "./domain/RestaurantService";
 import { mockList } from "./data/mockRestaurant";
 import { Template } from "./Template";
-import { BottomSheet } from "./until/ControlDom";
 
 class App {
   #app;
 
   constructor() {
     this.#app = document.querySelector("#app") as HTMLElement;
-    RestaurantList.list = mockList;
-    RestaurantList.settingList();
+    RestaurantService.list = mockList;
+    RestaurantService.settingList();
+    localStorage.clear();
   }
 
   render() {
     this.#app.innerHTML = `
-      ${Template.mainHeader()}
+      ${Template.mainHeader(BottomSheetOpenButton.template())}
       <main>
         <section class="restaurant-filter-container">
-          ${Template.filterButton}
-          ${Template.sortByButton}
+          ${FilterSortButton.template()}
         </section>
         <section class="restaurant-list-container">
-          ${Template.restaurantList(RestaurantList.list)}
+          ${Template.restaurantList(RestaurantService.list)}
         </section>
-
-        <div class="bottomSheet">
-          <div class="bottomSheet-backdrop"></div>
-          <div class="bottomSheet-container">
-            ${Template.bottomSheetHeader}
-            <form>
-              ${Template.categoryInput}
-              ${Template.nameInput}
-              ${Template.takeTimeInput}
-              ${Template.descriptionInput}
-              ${Template.linkInput}
-              ${Template.addButton}
-            </form>
-          </div>
-        </div>
       </main>
-    `;
+      ${BottomSheet.template()}`;
     this.setEvent();
   }
 
   setEvent() {
-    BottomSheetButton.setEvent();
-    addEventBottomSheet.addRestaurant();
-    addEventFilterSortButton();
+    BottomSheetOpenButton.openBottomSheet();
+    BottomSheet.addRestaurant();
+    FilterSortButton.FilterSort();
   }
 }
